Add: all updated js file and get rid of older ones.
